+++ src/ScottPlot5/ScottPlot5/Rendering/RenderActions/InvokePreRenderEvent.cs
@@ -1,0 +1,9 @@
+﻿namespace ScottPlot.Rendering.RenderActions;
+
+public class InvokePreRenderEvent : IRenderAction
+{
+    public void Render(RenderPack rp)
+    {
+        rp.Plot.RenderManager.RenderStarting.Invoke(this, rp);
+    }
+}