+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * AxisLine: Improve rendering and simplify API by exposing common properties (#3060, #3056)
 * Legend: `Alignment` has been renamed to `Location` (#3059)
 * Box: Refactored box plot API to favor simplicity and user customization (#3072)
+* Rendering: Added a pre-render event to `RenderManager` to allow modification of plottable properties (#3077) _Thanks GooBad_
 
 ## ScottPlot 4.1.70 (in development)
 * Population Plot: Improved performance for populations with curves that run off the screen (#3054) _Thanks @Em3a-c and @cornford_